SKALE-3030 Move config utils to helper
