+++ core/schains/cleaner.py
@@ -26,7 +26,7 @@ from multiprocessing import Process
 from core.schains.checks import SChainChecks
 from core.schains.helper import get_schain_dir_path
 from core.schains.runner import get_container_name, check_container_exit
-from core.schains.config_builder.utils import get_allowed_endpoints
+from core.schains.config_builder.helper import get_allowed_endpoints
 
 from sgx import SgxClient
